Changing the keys to a generic name for both prod and dev.
Using a key ID name rather than a UUID will allow us to keep one
Jenkinsfile for both prod and dev with just the choice param being
the switch. In future we can change this to a global var.

Needs further change to ensure the private key is gotten from the AWS
param store rather than placed in Jenkins manually as it currently has.